Seem to have fixed part of the crashing issue with logs and bit overwrites. Told VS to bundle in the runtime files needed. Runs on the remote server now.

+++ HK47/HK47/Move.cpp
@@ -55,7 +55,7 @@ void Move::moveToLocation(char row, char col)
 
 uint64_t Move::convertColRowToBitBoard(char col, char row)
 {
-	uint64_t thisLoc = 1;
+	uint64_t thisLoc = (uint64_t)1;
 
 	//these are the bit offsets we're going to use
 	//the row is counted numerically, so subtract 49.
